some changes

+++ driver_service/src/main/java/com/internship/driver_service/repo/RateRepo.java
@@ -5,10 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.util.List;
+import java.util.Map;
+
 public interface RateRepo extends JpaRepository<Rate, Long> {
     @Query("select FLOOR(AVG(r.value)) " +
             "from Rate r " +
             "where r.driver.profileId=:profileId ")
-    Byte findDriverRatingByProfileId(@Param("profileId") Long profileId);
-
+    Integer findDriverRatingByProfileId(@Param("profileId") Long profileId);
 }
